Create Gallery Image model
Created the Gallery Image model which allows the owner/admin to upload images to the database so that they can be used in the gallery page.

When the image is uploaded, it is resized to a more manageable size (300 x 500) or as close to that as possible. This is to reduce the size of the images which also means the loading times are faster.