Create Review model
Created the Review model which allows customers to create reviews and have them displayed on the review page.

The name field has a default in the event that the customer does not supply a name due to anonymity.

The date_added field is automatically set when the review is created.